+++ widgets/cSelectBox.h
@@ -17,7 +17,7 @@ public:
   virtual ~cSelectBox() {}
 
   //{{{
-  void onDown (int16_t x, int16_t y) {
+  void onDown (float x, float y) {
     cWidget::pressed (x, y);
     if (mMyValue != mValue) {
       mValue = mMyValue;
